atualizando url front

+++ db/migrations/20251219192422_create-meals.ts
@@ -14,7 +14,7 @@ export async function up(knex: Knex): Promise<void> {
     table.string('name').notNullable()
     table.text('description').notNullable()
     table.boolean('is_on_diet').notNullable()
-    table.integer('date').notNullable()
+    table.bigInteger('date').notNullable() 
 
     table
       .timestamp('created_at')
